styling(frontend): Simplifying the project after merge and restructure the folders. UT-188
