+++ frontend/src/admin/admin-page.scss
@@ -6,4 +6,10 @@
   grid-row-gap: 24px;
   height: 100vh;
   width: 100vw;
+  padding: 40px;
+  background-color: #171717;
+  color: #FFFFFF;
+  font-family: Helvetica Neue, sans-serif;
+  font-size: 16px;
+  box-sizing: border-box;
 }
