+++ frontend/src/admin/components/dropdown-menu-icon/dropdown-menu-icon.tsx
@@ -1,6 +1,6 @@
 import React from "react";
 
-export default function MenuIcon() {
+export default function DropdownMenuIcon() {
     return (
         <svg xmlns="http://www.w3.org/2000/svg" width="32" height="32" viewBox="0 0 32 32">
             <g id="Group_71" data-name="Group 71" transform="translate(-448 -83)">
