+++ frontend/src/admin/components/dropdown-menu-item/dropdown-menu-item.scss
@@ -1,0 +1,28 @@
+@import "../../../styles/main";
+
+p {
+  color: $contrast-color;
+  height: 30px;
+  padding: 16px;
+  border-bottom: 1px solid $border-color;
+  line-height: 2;
+  text-align: left;
+
+  &:last-child {
+    border-bottom: none;
+  }
+
+  &:hover {
+    cursor: pointer;
+    background: lighten($card-color, 20%);
+    transition:  0.3s ease-in-out;
+  }
+
+  &:hover:first-child {
+    border-radius: 5px 5px 0 0 ;
+  }
+
+  &:hover:last-child {
+    border-radius: 0 0 5px 5px ;
+  }
+}
