+++ frontend/src/admin/components/dropdown-menu/dropdown-menu.scss
@@ -1,0 +1,35 @@
+@import "../../../styles/main";
+
+.c-dropdown-menu {
+  position: relative;
+
+  button {
+    border: none;
+    margin: 0;
+    padding: 0;
+  }
+
+  .c-dropdown-menu-items {
+    display: none
+  }
+
+  &--active {
+    position: relative;
+
+
+    button {
+      border: none;
+      margin: 0;
+      padding: 0;
+    }
+
+    .c-dropdown-menu-items {
+      position: absolute;
+      display: block;
+      background-color: $card-color;
+      border-radius: 5px;
+      min-width: 150px;
+      box-shadow: 0 1px 8px rgba(0, 0, 0, 0.3);
+    }
+  }
+}
