+++ frontend/src/admin/components/restart-server-button/restart-server-button.scss
@@ -4,5 +4,5 @@
   background-color: $alert-color;
   border: 1px solid $alert-color;
   margin: 0;
-  grid-area: 5 / 1 / 6 / 2;
+  grid-area: 6 / 1 / 6 / 2;
 }
