+++ frontend/src/admin/components/restart-server-button/restart-server-button.tsx
