+++ frontend/src/admin/components/user-list-item/user-list-item.scss
