--- frontend/src/admin/components/user-list/icon-button.scss
@@ -1,7 +1,0 @@
-.c-icon-button {
-  background: none;
-  border: none;
-  cursor: pointer;
-  padding: 0;
-  width: auto;
-}
