--- frontend/src/admin/components/user-list/icon-button.tsx
@@ -1,11 +1,0 @@
-import React from "react";
-import MenuIcon from "./menu-icon";
-import "./icon-button.scss";
-
-export default function IconButton() {
-    return (
-        <button className="c-icon-button">
-            <MenuIcon/>
-        </button>
-    )
-}
